requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
blosc=1.21.3=h6c2663c_0
bottleneck=1.3.5=py37h080aedc_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotlipy=0.7.0=py37h2bbff1b_1003
bzip2=1.0.8=he774522_0
ca-certificates=2022.12.7=h5b45459_0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.1=py37h2bbff1b_3
cfitsio=3.470=h2bbff1b_7
charls=2.2.0=h6c2663c_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cloudpickle=2.0.0=pyhd3eb1b0_0
colorama=0.4.6=py37haa95532_0
cryptography=38.0.4=py37h21b164f_0
cuda=11.7.1=0
cuda-cccl=11.7.91=0
cuda-command-line-tools=11.7.1=0
cuda-compiler=11.7.1=0
cuda-cudart=11.7.99=0
cuda-cudart-dev=11.7.99=0
cuda-cuobjdump=11.7.91=0
cuda-cupti=11.7.101=0
cuda-cuxxfilt=11.7.91=0
cuda-demo-suite=12.0.140=0
cuda-documentation=12.0.140=0
cuda-libraries=11.7.1=0
cuda-libraries-dev=11.7.1=0
cuda-memcheck=11.8.86=0
cuda-nsight-compute=12.0.1=0
cuda-nvcc=11.7.99=0
cuda-nvdisasm=12.0.140=0
cuda-nvml-dev=11.7.91=0
cuda-nvprof=12.0.146=0
cuda-nvprune=11.7.91=0
cuda-nvrtc=11.7.99=0
cuda-nvrtc-dev=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-nvvp=12.0.146=0
cuda-runtime=11.7.1=0
cuda-sanitizer-api=12.0.140=0
cuda-toolkit=11.7.1=0
cuda-tools=11.7.1=0
cuda-visual-tools=11.7.1=0
cudatoolkit=10.1.243=h74a9793_0
cycler=0.11.0=pyhd3eb1b0_0
cytoolz=0.12.0=py37h2bbff1b_0
dask-core=2021.10.0=pyhd3eb1b0_0
fftw=3.3.9=h2bbff1b_1
flit-core=3.6.0=pyhd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=ha860e81_0
fsspec=2022.11.0=py37haa95532_0
giflib=5.2.1=h8cc25b3_3
glib=2.69.1=h5dc1a3c_2
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
idna=3.4=py37haa95532_0
imagecodecs=2021.8.26=py37hc0a7faf_1
imageio=2.19.3=py37haa95532_0
intel-openmp=2021.4.0=haa95532_3556
joblib=1.1.1=py37haa95532_0
jpeg=9e=h2bbff1b_0
kiwisolver=1.4.4=py37hd77b12b_0
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libaec=1.0.4=h33f27b4_1
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=12.0.0=default_h627e005_2
libcublas=11.10.3.66=0
libcublas-dev=11.10.3.66=0
libcufft=10.7.2.124=0
libcufft-dev=10.7.2.124=0
libcurand=10.3.1.124=0
libcurand-dev=10.3.1.124=0
libcusolver=11.4.0.1=0
libcusolver-dev=11.4.0.1=0
libcusparse=11.7.4.91=0
libcusparse-dev=11.7.4.91=0
libdeflate=1.8=h2bbff1b_5
libffi=3.4.2=hd77b12b_6
libiconv=1.16=h2bbff1b_2
libnpp=11.7.4.75=0
libnpp-dev=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libnvjpeg-dev=11.8.0.2=0
libogg=1.3.5=h2bbff1b_1
libpng=1.6.37=h2a8f88b_0
libtiff=4.5.0=h6c2663c_1
libuv=1.40.0=he774522_0
libvorbis=1.3.7=he774522_0
libwebp=1.2.4=h2bbff1b_0
libwebp-base=1.2.4=h2bbff1b_0
libxml2=2.9.14=h0ad7f3c_0
libxslt=1.1.35=h2bbff1b_0
libzopfli=1.0.3=ha925a31_0
locket=1.0.0=py37haa95532_0
lz4-c=1.9.4=h2bbff1b_0
matplotlib=3.5.2=py37haa95532_0
matplotlib-base=3.5.2=py37hd77b12b_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py37h2bbff1b_0
mkl_fft=1.3.1=py37h277e83a_0
mkl_random=1.2.2=py37hf11a4ad_0
munkres=1.1.4=py_0
networkx=2.6.3=pyhd3eb1b0_0
nsight-compute=2022.4.1.6=0
numexpr=2.8.4=py37h5b0cc5e_0
numpy=1.21.5=py37h7a0a035_3
numpy-base=1.21.5=py37hca35cd5_3
openjpeg=2.4.0=h4fc8c34_0
openssl=1.1.1t=h2bbff1b_0
packaging=22.0=py37haa95532_0
pandas=1.3.5=py37h6214cd6_0
partd=1.2.0=pyhd3eb1b0_1
pcre=8.45=hd77b12b_0
pickle5=0.0.11=py37h2bbff1b_0
pillow=9.3.0=py37hd77b12b_2
pip=22.3.1=py37haa95532_0
ply=3.11=py37_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py37haa95532_0
pyqt=5.15.7=py37hd77b12b_0
pyqt5-sip=12.11.0=py37hd77b12b_0
pysocks=1.7.1=py37_1
python=3.7.16=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch-cuda=11.7=h67b0de4_1
pytorch-mutex=1.0=cpu
pytz=2022.7=py37haa95532_0
pywavelets=1.3.0=py37h2bbff1b_0
pyyaml=6.0=py37h2bbff1b_1
qt-main=5.15.2=he8e5bd7_7
qt-webengine=5.15.9=hb9a9bb5_5
qtwebkit=5.212=h3ad3cdb_4
requests=2.28.1=py37haa95532_0
scikit-image=0.19.3=py37hd77b12b_1
scikit-learn=0.20.2=pypi_0
scipy=1.7.3=py37h7a0a035_2
setuptools=65.6.3=py37haa95532_0
sip=6.6.2=py37hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=h6c2663c_0
sqlite=3.40.1=h2bbff1b_0
tifffile=2021.7.2=pyhd3eb1b0_2
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.12.0=py37haa95532_0
torch=1.13.1+cu117=pypi_0
torchaudio=0.13.1=py37_cpu
torchvision=0.14.1=py37_cpu
tornado=6.2=py37h2bbff1b_0
tqdm=4.64.1=py37haa95532_0
typing_extensions=4.4.0=py37haa95532_0
urllib3=1.26.14=py37haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.38.4=py37haa95532_0
win_inet_pton=1.1.0=py37haa95532_0
wincertstore=0.2=py37haa95532_2
xz=5.2.10=h8cc25b3_1
yaml=0.2.5=he774522_0
zfp=0.5.5=hd77b12b_6
zlib=1.2.13=h8cc25b3_0
zstd=1.5.2=h19a0ad4_0
